refactor: 기존 TestController 의 이름 변경 및 테스트용 Controller 구현체 작성
