+++ mvc/src/test/java/samples/AnnotationTestController.java
@@ -1,0 +1,33 @@
+package samples;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.view.JspView;
+import nextstep.mvc.view.ModelAndView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.support.RequestMethod;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+@Controller
+public class AnnotationTestController {
+
+    private static final Logger log = LoggerFactory.getLogger(AnnotationTestController.class);
+
+    @RequestMapping(value = "/get-test", method = RequestMethod.GET)
+    public ModelAndView findUserId(HttpServletRequest request, HttpServletResponse response) {
+        log.info("test controller get method");
+        final ModelAndView modelAndView = new ModelAndView(new JspView(""));
+        modelAndView.addObject("id", request.getAttribute("id"));
+        return modelAndView;
+    }
+
+    @RequestMapping(value = "/post-test", method = RequestMethod.POST)
+    public ModelAndView save(HttpServletRequest request, HttpServletResponse response) {
+        log.info("test controller post method");
+        final ModelAndView modelAndView = new ModelAndView(new JspView(""));
+        modelAndView.addObject("id", request.getAttribute("id"));
+        return modelAndView;
+    }
+}
